// src/websocket/enum.ts
export enum SocketEvents {
  CONNECTION = "connection",
  ADD_NEW_USER = "addNewUser",
  SEND_MESSAGE = "sendMessage",
  DISCONNECT = "disconnect",
  NEW_CLIENT_CHAT = "newClientChat",
  DISCONNECT_CLIENT = "disconnectClient",
  ONLINE_USERS = "onlineUsers",
  GET_MESSAGE = "getMessage",
  NEW_USER_CHAT = "newUserChat"
}